+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/member/api/MemberApiController.java
@@ -2,7 +2,6 @@ package site.timecapsulearchive.core.domain.member.api;
 
 import jakarta.persistence.Access;
 import jakarta.validation.Valid;
-import java.time.ZonedDateTime;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
@@ -13,7 +12,6 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import site.timecapsulearchive.core.domain.member.data.dto.MemberDetailDto;
 import site.timecapsulearchive.core.domain.member.data.reqeust.CheckEmailDuplicationRequest;
@@ -22,7 +20,6 @@ import site.timecapsulearchive.core.domain.member.data.reqeust.UpdateFCMTokenReq
 import site.timecapsulearchive.core.domain.member.data.reqeust.UpdateNotificationEnabledRequest;
 import site.timecapsulearchive.core.domain.member.data.response.CheckEmailDuplicationResponse;
 import site.timecapsulearchive.core.domain.member.data.response.MemberDetailResponse;
-import site.timecapsulearchive.core.domain.member.data.response.MemberNotificationSliceResponse;
 import site.timecapsulearchive.core.domain.member.data.response.MemberNotificationStatusResponse;
 import site.timecapsulearchive.core.domain.member.data.response.MemberStatusResponse;
 import site.timecapsulearchive.core.domain.member.facade.MemberFacade;
@@ -106,21 +103,6 @@ public class MemberApiController implements MemberApi {
             ApiSpec.success(
                 SuccessCode.SUCCESS,
                 memberService.checkNotificationStatus(memberId)
-            )
-        );
-    }
-
-    @Override
-    @GetMapping(value = "/notifications")
-    public ResponseEntity<ApiSpec<MemberNotificationSliceResponse>> getMemberNotifications(
-        @AuthenticationPrincipal final Long memberId,
-        @RequestParam(defaultValue = "20", value = "size") final int size,
-        @RequestParam(value = "created_at") final ZonedDateTime createdAt
-    ) {
-        return ResponseEntity.ok(
-            ApiSpec.success(
-                SuccessCode.SUCCESS,
-                memberService.findNotificationSliceByMemberId(memberId, size, createdAt)
             )
         );
     }
