--- backend/core/src/main/java/site/timecapsulearchive/core/domain/member/data/dto/MemberNotificationDto.java
@@ -1,16 +1,0 @@
-package site.timecapsulearchive.core.domain.member.data.dto;
-
-import java.time.ZonedDateTime;
-import site.timecapsulearchive.core.domain.member.entity.CategoryName;
-import site.timecapsulearchive.core.domain.member.entity.NotificationStatus;
-
-public record MemberNotificationDto(
-    String title,
-    String text,
-    ZonedDateTime createdAt,
-    String imageUrl,
-    CategoryName categoryName,
-    NotificationStatus status
-) {
-
-}
