+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/member/data/mapper/MemberMapper.java
@@ -5,9 +5,9 @@ import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import site.timecapsulearchive.core.domain.member.data.dto.MemberNotificationDto;
-import site.timecapsulearchive.core.domain.member.data.response.MemberNotificationResponse;
-import site.timecapsulearchive.core.domain.member.data.response.MemberNotificationSliceResponse;
+import site.timecapsulearchive.core.domain.notification.data.dto.MemberNotificationDto;
+import site.timecapsulearchive.core.domain.notification.data.response.MemberNotificationResponse;
+import site.timecapsulearchive.core.domain.notification.data.response.MemberNotificationSliceResponse;
 import site.timecapsulearchive.core.domain.member.entity.Member;
 import site.timecapsulearchive.core.domain.member.entity.SocialType;
 import site.timecapsulearchive.core.global.security.oauth.dto.OAuth2UserInfo;
