--- backend/core/src/main/java/site/timecapsulearchive/core/domain/member/entity/NotificationStatus.java
@@ -1,5 +1,0 @@
-package site.timecapsulearchive.core.domain.member.entity;
-
-public enum NotificationStatus {
-    SUCCESS, FAIL
-}
