+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/member/repository/MemberQueryRepositoryImpl.java
@@ -2,25 +2,18 @@ package site.timecapsulearchive.core.domain.member.repository;
 
 import static site.timecapsulearchive.core.domain.friend.entity.QMemberFriend.memberFriend;
 import static site.timecapsulearchive.core.domain.member.entity.QMember.member;
-import static site.timecapsulearchive.core.domain.member.entity.QNotification.notification;
-import static site.timecapsulearchive.core.domain.member.entity.QNotificationCategory.notificationCategory;
 import static site.timecapsulearchive.core.domain.member_group.entity.QMemberGroup.memberGroup;
 
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.core.types.dsl.NumberExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.time.ZonedDateTime;
 import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Slice;
-import org.springframework.data.domain.SliceImpl;
 import org.springframework.stereotype.Repository;
 import site.timecapsulearchive.core.domain.member.data.dto.EmailVerifiedCheckDto;
 import site.timecapsulearchive.core.domain.member.data.dto.MemberDetailDto;
-import site.timecapsulearchive.core.domain.member.data.dto.MemberNotificationDto;
 import site.timecapsulearchive.core.domain.member.data.dto.VerifiedCheckDto;
 import site.timecapsulearchive.core.domain.member.entity.SocialType;
 
@@ -86,49 +79,6 @@ public class MemberQueryRepositoryImpl implements MemberQueryRepository {
 
     private NumberExpression<Long> countDistinct(final NumberExpression<Long> expression) {
         return Expressions.numberTemplate(Long.class, "COUNT(DISTINCT {0})", expression);
-    }
-
-    @Override
-    public Slice<MemberNotificationDto> findNotificationSliceByMemberId(
-        final Long memberId,
-        final int size,
-        final ZonedDateTime createdAt
-    ) {
-        final List<MemberNotificationDto> notifications = findMemberNotificationDtos(
-            memberId, size, createdAt);
-
-        final boolean hasNext = notifications.size() > size;
-        if (hasNext) {
-            notifications.remove(size);
-        }
-
-        return new SliceImpl<>(notifications, Pageable.ofSize(size), hasNext);
-    }
-
-    @Override
-    public List<MemberNotificationDto> findMemberNotificationDtos(
-        final Long memberId,
-        final int size,
-        final ZonedDateTime createdAt
-    ) {
-        return query
-            .select(
-                Projections.constructor(
-                    MemberNotificationDto.class,
-                    notification.title,
-                    notification.text,
-                    notification.createdAt,
-                    notification.imageUrl,
-                    notificationCategory.categoryName,
-                    notification.status
-                )
-            )
-            .from(notification)
-            .join(notification.notificationCategory, notificationCategory)
-            .where(notification.createdAt.lt(createdAt).and(notification.member.id.eq(memberId)))
-            .orderBy(notification.id.desc())
-            .limit(size + 1)
-            .fetch();
     }
 
     @Override
