+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/notification/entity/CategoryName.java
@@ -1,4 +1,4 @@
-package site.timecapsulearchive.core.domain.member.entity;
+package site.timecapsulearchive.core.domain.notification.entity;
 
 public enum CategoryName {
     CAPSULE_SKIN, FRIEND_REQUEST, FRIEND_ACCEPT, GROUP_INVITE, GROUP_ACCEPT
