+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/notification/entity/Notification.java
@@ -1,4 +1,4 @@
-package site.timecapsulearchive.core.domain.member.entity;
+package site.timecapsulearchive.core.domain.notification.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -16,6 +16,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
+import site.timecapsulearchive.core.domain.member.entity.Member;
 import site.timecapsulearchive.core.global.entity.BaseEntity;
 
 @Entity
