+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/notification/entity/NotificationStatus.java
@@ -1,0 +1,5 @@
+package site.timecapsulearchive.core.domain.notification.entity;
+
+public enum NotificationStatus {
+    SUCCESS, FAIL
+}
