+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/notification/repository/NotificationRepository.java
@@ -1,13 +1,14 @@
-package site.timecapsulearchive.core.domain.member.repository;
+package site.timecapsulearchive.core.domain.notification.repository;
 
 import java.time.ZonedDateTime;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.Repository;
 import org.springframework.data.repository.query.Param;
-import site.timecapsulearchive.core.domain.member.entity.Notification;
+import site.timecapsulearchive.core.domain.notification.entity.Notification;
 
-public interface NotificationRepository extends Repository<Notification, Long> {
+public interface NotificationRepository extends Repository<Notification, Long>,
+    NotificationQueryRepository {
 
     @Query("UPDATE Notification n SET n.deletedAt = :deletedAt WHERE n.member.id = :memberId")
     @Modifying
