+++ backend/core/src/main/java/site/timecapsulearchive/core/infra/queue/data/dto/GroupAcceptNotificationDto.java
@@ -1,7 +1,7 @@
 package site.timecapsulearchive.core.infra.queue.data.dto;
 
 import lombok.Builder;
-import site.timecapsulearchive.core.domain.member.entity.NotificationStatus;
+import site.timecapsulearchive.core.domain.notification.entity.NotificationStatus;
 
 @Builder
 public record GroupAcceptNotificationDto(
